make code a bit better

+++ src/Button.cpp
@@ -2,7 +2,7 @@
 
 #include "raygui.h"
 
-Button::Button(const int x, const int y, const float width, const float height, const std::string &text, const std::string& tooltip, const Color color) {
+Button::Button(const float x, const float y, const float width, const float height, const std::string &text, const std::string& tooltip, const Color color) {
 	this->x = x;
 	this->y = y;
 	this->width = width;
@@ -20,10 +20,10 @@ Button::~Button() = default;
 
 
 void Button::draw() {
-	const auto rect = Rectangle{.x = static_cast<float>(x), .y = static_cast<float>(y), .width = width, .height = height};
+	const auto rect = Rectangle{.x = x, .y = y, .width = width, .height = height};
 	pressed = GuiButton(rect, text.c_str());
 	if (icon != -1) {
-		GuiDrawIcon(icon, x, y, static_cast<int>(width) / 16, color);
+		GuiDrawIcon(icon, static_cast<int>(x), static_cast<int>(y), static_cast<int>(width) / 16, color);
 	}
 	if (CheckCollisionPointRec(GetMousePosition(), rect)) {
 		if (tooltip.empty()) {
